Renaming and Moving Tests To BrowseLanguagesTests
