+++ src/test/java/tests/browse_languages/languages/LanguageCSharpTest.java
@@ -1,10 +1,10 @@
-package tests._to_do;
+package tests.browse_languages.languages;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import runner.BaseTest;
 
-public class CountCSharpLanguageVersionsTest extends BaseTest {
+public class LanguageCSharpTest extends BaseTest {
 
     @Test
     public void testNumberOfCSharpLanguageVersions() {
